refactor(edgeless): remove hard writing edgeless tool types (#7296)
# What changed?

This PR changes the way declarations of edgeless tool.

## Before
```ts
export type MyCustomEdgelessTool = {
    type: "my-custom-tool";
    // ...
};

export class MyCustomEdgelessToolController extends EdgelessToolController<MyCustomEdgelessTool> {
    // ...
};

// Update the type `EdgelessTool`
export type EdgelessTool =
    DefaultTool |
    // ...
    MyCustomEdgelessTool;
```

## After
```ts
export type MyCustomEdgelessTool = {
    type: "my-custom-tool";
    // ...
};

export class MyCustomEdgelessToolController extends EdgelessToolController<MyCustomEdgelessTool> {
    // ...
};

// Non-invasive update type `EdgelessTool`
declare global {
  namespace BlockSuite {
    interface EdgelessToolMap {
      "my-custom-tool": MyCustomEdgelessToolController;
    }
  }
}
```
The implementation is in https://github.com/toeverything/blocksuite/pull/7296/files#diff-20771409cc72bce6e327202907e2dbd6fef600e2786129af4834017a0aca6de6
