+++ packages/blocks/src/root-block/edgeless/components/toolbar/eraser/eraser-tool-button.ts
@@ -5,7 +5,7 @@ import {
   EdgelessEraserDarkIcon,
   EdgelessEraserLightIcon,
 } from '../../../../../_common/icons/index.js';
-import type { EdgelessTool } from '../../../../../_common/utils/index.js';
+import type { EdgelessTool } from '../../../types.js';
 import { getTooltipWithShortcut } from '../../utils.js';
 import { EdgelessToolbarToolMixin } from '../mixins/tool.mixin.js';
