+++ packages/blocks/src/root-block/edgeless/components/toolbar/frame/frame-menu.ts
@@ -2,7 +2,7 @@ import { css, html, LitElement } from 'lit';
 import { customElement } from 'lit/decorators.js';
 import { repeat } from 'lit/directives/repeat.js';
 
-import type { EdgelessTool } from '../../../../../_common/utils/index.js';
+import type { EdgelessTool } from '../../../types.js';
 import { EdgelessToolbarToolMixin } from '../mixins/tool.mixin.js';
 import { FrameConfig } from './config.js';
 import { createFrame } from './service.js';
