+++ packages/blocks/src/root-block/edgeless/components/toolbar/mixins/tool.mixin.ts
@@ -5,8 +5,8 @@ import { cssVar } from '@toeverything/theme';
 import type { LitElement } from 'lit';
 import { property, state } from 'lit/decorators.js';
 
-import type { EdgelessTool } from '../../../../../_common/types.js';
 import type { EdgelessRootBlockComponent } from '../../../edgeless-root-block.js';
+import type { EdgelessTool } from '../../../types.js';
 import { createPopper, type MenuPopper } from '../common/create-popper.js';
 import { edgelessToolbarThemeContext } from '../context.js';
