+++ packages/blocks/src/root-block/edgeless/controllers/tools/pan-tool.ts
@@ -1,8 +1,12 @@
 import type { PointerEventState } from '@blocksuite/block-std';
 import { noop } from '@blocksuite/global/utils';
 
-import type { PanTool } from '../../../../_common/utils/index.js';
-import { EdgelessToolController } from './index.js';
+import { EdgelessToolController } from './edgeless-tool.js';
+
+type PanTool = {
+  type: 'pan';
+  panning: boolean;
+};
 
 export class PanToolController extends EdgelessToolController<PanTool> {
   private _lastPoint: [number, number] | null = null;
@@ -84,5 +88,13 @@ export class PanToolController extends EdgelessToolController<PanTool> {
 
   afterModeSwitch() {
     noop();
+  }
+}
+
+declare global {
+  namespace BlockSuite {
+    interface EdgelessToolMap {
+      pan: PanToolController;
+    }
   }
 }
