+++ packages/blocks/src/root-block/edgeless/index.ts
@@ -1,0 +1,7 @@
+export { FramePreview } from './components/frame/frame-preview.js';
+export { EdgelessToolController } from './controllers/tools/index.js';
+export { EdgelessBlockModel } from './edgeless-block-model.js';
+export * from './edgeless-root-block.js';
+export { EdgelessRootService } from './edgeless-root-service.js';
+export * from './types.js';
+export { Viewport } from './utils/viewport.js';
