+++ packages/blocks/src/root-block/edgeless/utils/query.ts
@@ -1,10 +1,7 @@
 import type { BlockModel } from '@blocksuite/store';
 
 import type { EmbedBlockModel } from '../../../_common/embed-block-helper/embed-block-model.js';
-import type {
-  Connectable,
-  EdgelessTool,
-} from '../../../_common/utils/index.js';
+import type { Connectable } from '../../../_common/utils/index.js';
 import type { AttachmentBlockModel } from '../../../attachment-block/index.js';
 import type { BookmarkBlockModel } from '../../../bookmark-block/bookmark-model.js';
 import type { EdgelessTextBlockModel } from '../../../edgeless-text/edgeless-text-model.js';
@@ -33,6 +30,7 @@ import {
   TextElementModel,
 } from '../../../surface-block/index.js';
 import type { EdgelessBlockModel } from '../edgeless-block-model.js';
+import type { EdgelessTool } from '../types.js';
 import { getElementsWithoutGroup } from './group.js';
 import type { Viewport } from './viewport.js';
