+++ packages/blocks/src/root-block/index.ts
@@ -5,12 +5,7 @@ import type { PageRootService } from './page/page-root-service.js';
 import type { RootBlockModel } from './root-model.js';
 
 export * from './clipboard/index.js';
-export { FramePreview } from './edgeless/components/frame/frame-preview.js';
-export { EdgelessBlockModel } from './edgeless/edgeless-block-model.js';
-export * from './edgeless/edgeless-root-block.js';
-export { EdgelessRootService } from './edgeless/edgeless-root-service.js';
-export * from './edgeless/types.js';
-export { Viewport } from './edgeless/utils/viewport.js';
+export * from './edgeless/index.js';
 export * from './page/page-root-block.js';
 export { PageRootService } from './page/page-root-service.js';
 export * from './preview/preview-root-block.js';
